customer-behavior-analytics-dashboard/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
